chore(merge): move duckdb helpers out of ee, delete dead rekeyMergeRows
The composition engine must not depend on EE-gated code paths, and two of
its helpers lived under src/ee for no reason beyond where they were first
needed — both were already imported by OSS code. They live in utils/duckdb
now, with the pre-aggregate services importing from there.

rekeyMergeRows had no callers: built for client-side rekeying, superseded
by the statement renaming its own columns to field ids.
