+++ packages/backend/src/utils/duckdb/duckdbSqlTables.ts
